Adding bones with invbind matrices is bad idea
Unity SkinnedMeshRenderer only stores bones that are active on the mesh.
Before, we generated skins with added bones in order to have proper tree
hierarchy with one single root bone. Dummy identity matrices were set as
invbindmatrices for these added bones.

If bone is added in skin1 with dummy identity invbind and used in skin2 with
real invbind, if skin1 is processed before skin2, it's bad (for skfb)
In this way, it's useless to specify root skeleton for skin